Add another ForwardDiff workaround for StaticArrays
See https://github.com/PALEOtoolkit/PALEOmodel.jl/issues/40

Add a workaround for integrateLocalIMEXEuler
